// src/app/Pages/types.ts
// export interface Product{
//     id: number;
//     name: string;
//     price: number;
//     image: string;
// }
export interface Product {
    id: number;    // Unique identifier for the product
    name: string;  // Name of the product
    price: number; // Price of the product
    image: string; // Image URL for the product
  }
  